refactor: extract a shared ChatColumnCell for chat rows
ChatMessageCell and ChatCashCardCell had drifted into an identical column +
receipt scaffold, leading/trailing alignment constraints, and a character-for-
character applyAlignment. Lift that into a ChatColumnCell base so each cell keeps
only its own content (a bubble or a card). Also drop ChatViewController's
sizeForItem override, which now only restated ChatLayout's .auto default.
